+++ README.md
@@ -1,4 +1,4 @@
-git## fitness_app
+## fitness_app
 Fitness app for tracking exercise progress.
 
 #Resources
