+++ app/baselineprofile/src/main/AndroidManifest.xml
@@ -1,1 +1,1 @@
-<manifest/>
+<manifest />
